modularize openmp and creates .gitignore
