+++ main.c
@@ -1,4 +1,5 @@
 #include "pthreads/ptmatrix.h"
+#include "openmp/ommatrix.h"
 #include "util.h"
 #include <string.h>
 
@@ -33,8 +34,8 @@ int main(int argc, char *argv[]) {
   // validates if a is able to be multiplied by b.
   if (matrix_mult_valid(a, b))
   {
-	  // Matrix_t c = matrix_mult_openmp(a, b, 4, 1);
-    Matrix_t c = matrix_mult_pthread(a, b, 4);
+	  Matrix_t c = matrix_mult_openmp(a, b, 4, 1);
+    // Matrix_t c = matrix_mult_pthread(a, b, 4);
     matrix_print_to_file(matrix_file_c, c);
     printf("Results written to %s\n", matrix_path_c);
     release_resources();
